Replace BreakpointMenu with CommonActions, which now handles more commands
