+++ src/gui/Src/Gui/CPUDisassembly.h
@@ -2,12 +2,12 @@
 #define CPUDISASSEMBLY_H
 
 #include "Disassembly.h"
-#include "BreakpointMenu.h"
 
 // Needed forward declaration for parent container class
 class CPUSideBar;
 class GotoDialog;
 class XrefBrowseDialog;
+class CommonActions;
 
 class CPUDisassembly : public Disassembly
 {
@@ -41,8 +41,8 @@ public slots:
     void gotoOriginSlot();
     void setLabelSlot();
     void setLabelAddressSlot();
-    void setCommentSlot();
-    void setBookmarkSlot();
+    //void setCommentSlot();
+    //void setBookmarkSlot();
     void toggleFunctionSlot();
     void toggleArgumentSlot();
     void addLoopSlot();
@@ -110,7 +110,7 @@ public slots:
     void createThreadSlot();
     void copyTokenTextSlot();
     void copyTokenValueSlot();
-    void followInMemoryMapSlot();
+    //void followInMemoryMapSlot();
     void downloadCurrentSymbolsSlot();
 
 protected:
@@ -169,7 +169,7 @@ private:
     MenuBuilder* mMenuBuilder;
     MenuBuilder* mHighlightMenuBuilder;
     bool mHighlightContextMenu = false;
-    BreakpointMenu* mBreakpointMenu;
+    CommonActions* mCommonActions;
 };
 
 #endif // CPUDISASSEMBLY_H
