+++ src/gui/Src/Gui/CPUDump.h
@@ -7,6 +7,7 @@
 class CPUMultiDump;
 class CPUDisassembly;
 class GotoDialog;
+class CommonActions;
 
 class CPUDump : public HexDump
 {
@@ -98,11 +99,11 @@ public slots:
     void findPattern();
     void copyFileOffsetSlot();
     void undoSelectionSlot();
-    void followStackSlot();
+    //void followStackSlot();
     void findReferencesSlot();
-    void followInDisasmSlot();
-    void followDataSlot();
-    void followDataDumpSlot();
+    //void followInDisasmSlot();
+    //void followDataSlot();
+    //void followDataDumpSlot();
 
     void watchSlot();
 
@@ -111,12 +112,13 @@ public slots:
     void followInDumpNSlot();
     void allocMemorySlot();
 
-    void followInMemoryMapSlot();
+    //void followInMemoryMapSlot();
     void headerButtonReleasedSlot(int colIndex);
     void asciiAddressDumpModeUpdatedSlot();
 
 private:
     MenuBuilder* mMenuBuilder;
+    CommonActions* mCommonActions;
 
     QMenu* mPluginMenu;
     QMenu* mFollowInDumpMenu;
