+++ src/x64dbg_translations.pro
@@ -217,7 +217,7 @@ SOURCES += \
     gui/Src/Gui/LocalVarsView.cpp \
     gui/Src/Gui/MessagesBreakpoints.cpp \
     gui/Src/Gui/AboutDialog.cpp \
-    gui/Src/Gui/BreakpointMenu.cpp \
+    gui/Src/Gui/CommonActions.cpp \
     gui/Src/Utils/SymbolAutoCompleteModel.cpp \
     gui/Src/Tracer/TraceBrowser.cpp \
     gui/Src/Tracer/TraceFileReader.cpp \
@@ -457,7 +457,7 @@ HEADERS += \
     gui/Src/Gui/LocalVarsView.h \
     gui/Src/Gui/MessagesBreakpoints.h \
     gui/Src/Gui/AboutDialog.h \
-    gui/Src/Gui/BreakpointMenu.h \
+    gui/Src/Gui/CommonActions.h \
     gui/Src/Utils/SymbolAutoCompleteModel.h \
     gui/Src/Tracer/TraceBrowser.h \
     gui/Src/Tracer/TraceFileReader.h \
